Move common completion methods to CompletionHelpers class
